SPP example initial commit
These two files will perform an SPP test.
The python script connects to the BT900 and opens an SPP pipe.
The script is currently failing because of the rfcomm linux tool used on
the python script.
If the BT900 is used with the android app Blueterm which emulates a
serial over Bluetooth classic, the test works.